rip out SqlTableContext, World is the knowledge source for everything, much better
