+++ allennlp/tests/data/dataset_readers/semantic_parsing/grammar_based_text2sql_test.py
@@ -18,6 +18,5 @@ class TestGrammarBasdText2SqlDatasetReader(AllenNlpTestCase):
 
         assert len(instances) == 5
 
-
         for instance in instances:
             print(instance)
